python/requirements.cu12.txt
pip>22.0.2
numpy>=1.0.4,<3.0
cloudpickle>=3.0,<4.0
pillow>=11.2
cupy-cuda12x>=12.2,<14.0
